requirements.txt
invoke >= 0.20
rellu >= 0.6
docutils >= 0.14
robotframework >= 3.1.1
robotframework-seleniumlibrary >= 3.3.1
selenium
selenium-datatable
msedge-selenium-tools
PyHamcrest